Propagate SQLite return codes to the developer
Haskell is a high concurrency language, if you use SQLite for
application data, you might need multiple threads to access the same
SQLite database, therefore, this database could be locked or busy.

This allows the developer to handle BUSY and/or LOCKED error code, and
wait for some transactions to finish when using SQLite is a
multi-threaded environment.